compbio_ml_projects/openvirotax/requirements.txt
absl-py==2.3.1
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array-api-compat==1.12.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
biom-format==2.1.17
biopython==1.85
black==25.1.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flake8==7.3.0
flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.2
fqdn==1.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.13
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
keras==3.11.3
kiwisolver==1.4.9
lark==1.2.2
libclang==18.1.1
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
ml-dtypes==0.5.3
mypy==1.17.1
mypy-extensions==1.1.0
namex==0.1.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.4.5
notebook-shim==0.2.4
numpy==2.3.2
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
-e file:///home/dev_ampofo/bioinformatics/compbio_ml_projects/pathogen_classifier
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==4.3.0
prometheus-client==0.22.1
prompt-toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
scikit-bio==0.7.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
send2trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
statsmodels==0.14.5
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing-extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
virtualenv==20.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
